+++ apina-core/src/test/kotlin/fi/evident/apina/spring/SpringModelReaderTest.kt
@@ -3,13 +3,19 @@ package fi.evident.apina.spring
 import fi.evident.apina.java.model.JavaModel
 import fi.evident.apina.java.reader.loadClassesFromInheritanceTree
 import fi.evident.apina.model.ApiDefinition
+import fi.evident.apina.model.Endpoint
+import fi.evident.apina.model.EndpointGroup
 import fi.evident.apina.model.HTTPMethod
 import fi.evident.apina.model.settings.TranslationSettings
 import fi.evident.apina.model.type.ApiType
 import org.junit.Test
+import org.springframework.http.HttpEntity
+import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import java.util.concurrent.Callable
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
@@ -38,10 +44,8 @@ class SpringModelReaderTest {
         val endpointGroup = model.endpointGroups.single()
         assertEquals("My", endpointGroup.name)
 
-        val foo = assertNotNull(endpointGroup.endpoints.find { it.name == "foo" }, "foo")
-        val bar = assertNotNull(endpointGroup.endpoints.find { it.name == "bar" }, "bar")
-        assertEquals(HTTPMethod.POST, foo.method)
-        assertEquals(HTTPMethod.GET, bar.method)
+        assertEquals(HTTPMethod.POST, endpointGroup.endpointByName(("foo")).method)
+        assertEquals(HTTPMethod.GET, endpointGroup.endpointByName(("bar")).method)
     }
 
     @Test
@@ -57,11 +61,42 @@ class SpringModelReaderTest {
             fun foo(): Foo<String> = error("")
         }
 
-        val model = readModel<ControllerWithVariables>()
+        val group = readModel<ControllerWithVariables>().endpointGroups.single()
 
-        val endpoint = model.endpointGroups.single().endpoints.single()
-        assertEquals(ApiType.Class("Foo"), assertNotNull(endpoint.responseBody))
+        assertEquals(ApiType.Class("Foo"), group.endpointByName("foo").responseBody)
     }
+
+    @Test
+    fun wrappedResultTypes() {
+
+        class Foo
+
+        @RestController
+        class MyController {
+
+            @Suppress("unused")
+            @RequestMapping("/foo")
+            fun foo(): ResponseEntity<Foo> = error("dummy")
+
+            @Suppress("unused")
+            @RequestMapping("/bar")
+            fun bar(): HttpEntity<Foo> = error("dummy")
+
+            @Suppress("unused")
+            @RequestMapping("/bar")
+            fun baz(): Callable<Foo> = error("dummy")
+        }
+
+        val group = readModel<MyController>().endpointGroups.single()
+
+        assertEquals(ApiType.Class("Foo"), group.endpointByName("foo").responseBody)
+        assertEquals(ApiType.Class("Foo"), group.endpointByName("bar").responseBody)
+        assertEquals(ApiType.Class("Foo"), group.endpointByName("baz").responseBody)
+    }
+
+
+    fun EndpointGroup.endpointByName(name: String): Endpoint =
+        assertNotNull(endpoints.find { it.name == name })
 
     private inline fun <reified T : Any> readModel(): ApiDefinition {
         val model = JavaModel()